+++ CHANGES.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `IScope.ContainedScopes` which lets you check scopes contained within another scope.
 - Added `IVariable.CanBeAccessedIn(IScope)` which lets you check if the variable can be accessed within a given scope.
 - Added `IScope.FindVariable(string)` which lets you try to find a variable in a scope or any of its parents.
+- Added `Script.RenameVariable(IVariable variable, string newName)` which lets you rename a variable in a new script instance.
 
 ### Changed
 - Renamed `IScope.Parent` to `IScope.ContainingScope` to be more consistent with `ContainedScopes`.